requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-auth==2.35.0
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
python-dotenv==1.0.1
